backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
langchain==0.3.1
langchain-openai==0.2.2
langchain-qdrant==0.1.4
langchain-community==0.3.1
python-dotenv==1.0.1
openai==1.51.0
pypdf==5.0.1
qdrant-client==1.11.3